Раздел: Общие изменения
Расщепил RecipePreview на набор подкомпонентов, хуков и утилит с отдельными стилями для повышения читаемости и поддержки.
Добавил типы и константы для работы с данными превью, вынес повторяющиеся вычисления в специализированные хуки.
Раздел: Багфиксы

Исправил setRecipeAction, отправляя строковые значения в payload согласно контракту API.
Поправил импорты в хуках превью, чтобы сборка проходила без ошибки TypeScript.
Раздел: Новые фичи

Добавил компактный вывод бейджей в виде чипов прямо в блоке категорий превью рецепта.
