+++ .gitignore
@@ -12,6 +12,7 @@ dist
 dist-ssr
 *.local
 api
+notes
 
 # Editor directories and files
 .vscode/*
