--- public/test.md
@@ -1,4 +1,0 @@
-- Давай сделаем краткий коммит на русском с git add . и сделаем пуш
-- Давай напишем для компонента подробный JSdoc
-- Давай сделаем новую ветку для фичи feature, имя текущая дата и время.
-- Давай переключимся на develop и сделаем pull
