+++ src/components/RecipePreview/components/BadgesChips/BadgesChips.module.css
@@ -1,0 +1,73 @@
+.section {
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacing-sm);
+}
+
+.title {
+  font-size: var(--font-size-xl);
+  font-weight: var(--font-weight-semibold);
+  color: var(--text-primary);
+  margin: 0;
+}
+
+.list {
+  display: flex;
+  flex-wrap: wrap;
+  gap: var(--spacing-sm);
+}
+
+.chip {
+  display: inline-flex;
+  align-items: center;
+  gap: var(--spacing-xs);
+  border: 1px solid transparent;
+  border-radius: var(--token-radius-full);
+  padding: 6px 12px;
+  min-height: 40px;
+  transition: transform 0.2s ease, box-shadow 0.2s ease;
+  box-shadow: var(--token-shadow-xs);
+}
+
+.chip:hover {
+  transform: translateY(-1px);
+  box-shadow: var(--token-shadow-sm);
+}
+
+.chipIcon {
+  width: 28px;
+  height: 28px;
+  border-radius: var(--token-radius-full);
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  font-size: var(--font-size-sm);
+  overflow: hidden;
+  color: var(--text-primary);
+}
+
+.chipIcon img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.chipContent {
+  display: inline-flex;
+  flex-direction: column;
+  gap: 4px;
+  max-width: 240px;
+}
+
+.chipTitle {
+  font-size: var(--font-size-sm);
+  font-weight: var(--font-weight-semibold);
+  color: var(--text-primary);
+  line-height: 1.2;
+}
+
+.chipDescription {
+  font-size: var(--font-size-xs);
+  color: var(--text-secondary);
+  line-height: 1.2;
+}
