+++ src/components/RecipePreview/components/CommentsSection/CommentsSection.module.css
@@ -1,0 +1,11 @@
+.section {
+  margin-top: 24px;
+}
+
+.listItem {
+  margin-bottom: 12px;
+}
+
+.title {
+  margin: 0 0 16px 0;
+}
