+++ src/components/RecipePreview/components/HeroSection/HeroSection.module.css
@@ -1,0 +1,3 @@
+.root {
+  width: 100%;
+}
