+++ src/components/RecipePreview/components/InfoGrid/InfoGrid.module.css
@@ -1,0 +1,35 @@
+.grid {
+  display: grid;
+  grid-template-columns: repeat(4, minmax(0, 1fr));
+  gap: var(--token-space-6);
+  margin: var(--token-space-6) 0;
+}
+
+.infoIcon {
+  width: 22px;
+  height: 22px;
+  color: var(--token-muted);
+}
+
+.difficultyBadge {
+  display: inline-flex;
+  align-items: center;
+  height: 32px;
+  padding: 0 var(--token-space-4);
+  border-radius: var(--token-radius-full);
+  font-weight: 700;
+}
+
+@media (max-width: 1100px) {
+  .grid {
+    grid-template-columns: repeat(2, minmax(0, 1fr));
+    gap: var(--token-space-6);
+  }
+}
+
+@media (max-width: 650px) {
+  .grid {
+    grid-template-columns: repeat(2, minmax(0, 1fr));
+    gap: var(--token-space-4);
+  }
+}
