+++ src/components/RecipePreview/components/MainSections/MainSections.module.css
@@ -1,0 +1,12 @@
+.contentGrid {
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacing-3xl);
+  margin-bottom: var(--spacing-3xl);
+}
+
+@media (max-width: 768px) {
+  .contentGrid {
+    gap: var(--spacing-2xl);
+  }
+}
