+++ src/components/RecipePreview/components/ReviewSection/ReviewSection.module.css
@@ -1,0 +1,48 @@
+.section {
+  background: var(--token-surface);
+  border-radius: var(--token-radius-lg);
+  padding: var(--token-space-6);
+  box-shadow: var(--token-shadow-sm);
+  margin: 0 0 var(--token-space-6) 0;
+}
+
+.title {
+  margin: 0 0 var(--token-space-4) 0;
+  color: var(--token-text);
+  font-size: var(--token-text-xl);
+  font-weight: 700;
+}
+
+.controls {
+  display: flex;
+  align-items: center;
+  gap: var(--token-space-4);
+  margin-bottom: var(--token-space-4);
+}
+
+.label {
+  color: var(--token-muted);
+}
+
+.field {
+  margin-bottom: var(--token-space-4);
+}
+
+.textarea {
+  width: 100%;
+  min-height: 110px;
+  resize: vertical;
+  border: none;
+  outline: none;
+  border-radius: var(--token-radius-md);
+  background: var(--token-surface);
+  color: var(--token-text);
+  box-shadow: inset 0 0 0 1px var(--token-border);
+  padding: 12px 14px;
+  box-sizing: border-box;
+}
+
+.actions {
+  display: flex;
+  justify-content: flex-end;
+}
