+++ src/components/RecipePreview/components/SocialActions/SocialActions.module.css
@@ -1,0 +1,5 @@
+.actions {
+  display: inline-flex;
+  align-items: center;
+  gap: 8px;
+}
